refactor serde crate to avoid name duplication
