+++ crates/gitbutler-project/src/project.rs
@@ -61,7 +61,7 @@ impl FetchResult {
 
 #[derive(Debug, Deserialize, Serialize, Copy, Clone)]
 pub struct CodePushState {
-    #[serde(with = "gitbutler_serde::serde::oid")]
+    #[serde(with = "gitbutler_serde::oid")]
     pub id: git2::Oid,
     pub timestamp: time::SystemTime,
 }
